core/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
docx==0.2.4
duckduckgo-search==2.8.5
filelock==3.11.0
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
httplib2==0.22.0
idna==3.4
loguru==0.6.0
lxml==4.9.2
mpmath==1.2.1
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
orjson==3.8.9
packaging==23.0
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
playsound==1.2.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4
striprtf==0.0.25
# Editable install with no version control (sympy==1.11.1)
-e /Library/Frameworks/Python.framework/Versions/3.11/lib/python3.11/site-packages
tiktoken==0.3.3
tqdm==4.64.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
websocket-client==1.5.1
yarl==1.8.2
